git-svn-id: https://subversion.imagemagick.org/subversion/ImageMagick/trunk@2088 aa41f4f7-0bf4-0310-aa73-e5a19afd5a74

+++ Magick++/lib/Magick++/STL.h
@@ -822,7 +822,7 @@ namespace Magick
     void operator()( Image &image_ ) const;
 
   private:
-    size_t _opacity;
+    unsigned int _opacity;
   };
 
   // Change color of opaque pixel to specified pen color.
